refactor: simplify runSearch signature to accept QueryRecord
Move QueryRecord interface to search.ts (avoiding circular dependency)
and re-export from searchPanel.ts. Change runSearch(db, query,
groupingMode, sortOptions) to runSearch(db, params: QueryRecord),
simplifying all 19 call sites across searchPanel.ts, index.ts, and
noteView.ts.
