+++ src/searchPanel.ts
@@ -5,7 +5,7 @@ import * as markdownItTaskLists from 'markdown-it-task-lists';
 import * as prism from './prism.js';
 import { TagSettings, getTagSettings, queryEnd, queryStart, getResultSettings, getStandardGroupingKeys } from './settings';
 import { escapeRegex, mapPrefixClass } from './utils';
-import { GroupedResult, Query, runSearch, sortResults } from './search';
+import { GroupedResult, Query, QueryRecord, runSearch, sortResults } from './search';
 import { noteIdRegex } from './parser';
 import { NoteDatabase, processNote } from './db';
 import { validatePanelMessage, ValidationError } from './validation';
@@ -46,27 +46,8 @@ export const REGEX = {
   coreMarker: /\u200B\u2061/g,
 };
 
-/**
- * Represents a search query configuration
- */
-export interface QueryRecord {
-  /** Array of query conditions grouped by AND/OR logic */
-  query: Query[][];
-  /** Text filter to apply to results */
-  filter: string;
-  /** How to display results in the note: 'false', 'list', 'table', or 'kanban' */
-  displayInNote: string;
-  /** Optional display settings */
-  options?: {
-    includeCols?: string;
-    excludeCols?: string;
-    sortBy?: string;
-    sortOrder?: string;
-    resultGrouping?: string;
-    resultToggle?: boolean;
-    limit?: number;
-  };
-}
+// QueryRecord is defined in search.ts and re-exported here for backward compatibility
+export { QueryRecord } from './search';
 
 /**
  * Interface for messages received from the search panel UI
@@ -274,7 +255,7 @@ async function processValidatedMessage(
     await updatePanelQuery(searchPanel, searchParams.query, searchParams.filter);
     await updatePanelSettings(searchPanel, searchParams);
 
-    const results = await runSearch(db, searchParams.query, searchParams.options?.resultGrouping, searchParams.options);
+    const results = await runSearch(db, searchParams);
     lastSearchResults = results; // Cache the results
     await updatePanelResults(searchPanel, results, searchParams.query, searchParams.options);
     await updatePanelNoteState(searchPanel, savedNoteState);
@@ -288,7 +269,7 @@ async function processValidatedMessage(
     }
 
     // Run the search and update results
-    const results = await runSearch(db, searchParams.query, searchParams.options?.resultGrouping, searchParams.options);
+    const results = await runSearch(db, searchParams);
     lastSearchResults = results; // Update cached results
     await updatePanelResults(searchPanel, results, searchParams.query, searchParams.options);
 
@@ -404,7 +385,7 @@ async function processValidatedMessage(
     await setCheckboxState(message, db, tagSettings);
 
     // update the search panel
-    const results = await runSearch(db, searchParams.query, searchParams.options?.resultGrouping, searchParams.options);
+    const results = await runSearch(db, searchParams);
     lastSearchResults = results; // Cache the results
     
     // Apply sorting to maintain current sort order
@@ -421,7 +402,7 @@ async function processValidatedMessage(
 
     // update the search panel
     await updatePanelTagData(searchPanel, db);
-    const results = await runSearch(db, searchParams.query, searchParams.options?.resultGrouping, searchParams.options);
+    const results = await runSearch(db, searchParams);
     lastSearchResults = results; // Cache the results
     
     // Apply sorting to maintain current sort order
@@ -441,7 +422,7 @@ async function processValidatedMessage(
 
     // update the search panel
     await updatePanelTagData(searchPanel, db);
-    const results = await runSearch(db, searchParams.query, searchParams.options?.resultGrouping, searchParams.options);
+    const results = await runSearch(db, searchParams);
     lastSearchResults = results; // Cache the results
     
     // Apply sorting to maintain current sort order
@@ -456,7 +437,7 @@ async function processValidatedMessage(
     await addTagToText(message, db, tagSettings);
 
     // update the search panel
-    const results = await runSearch(db, searchParams.query, searchParams.options?.resultGrouping, searchParams.options);
+    const results = await runSearch(db, searchParams);
     lastSearchResults = results; // Cache the results
     
     // Apply sorting to maintain current sort order
@@ -500,7 +481,7 @@ async function processValidatedMessage(
 
           // Re-run search with new grouping
           if (lastSearchResults && lastSearchResults.length > 0 && searchParams.query && searchParams.query.length > 0) {
-            const results = await runSearch(db, searchParams.query, message.value, searchParams.options);
+            const results = await runSearch(db, searchParams);
             lastSearchResults = results; // Update cached results
             await updatePanelResults(searchPanel, results, searchParams.query, searchParams.options);
           }
@@ -579,7 +560,7 @@ async function processValidatedMessage(
           await updatePanelSettings(searchPanel, searchParams);
 
           // Run search with the loaded query
-          const results = await runSearch(db, searchParams.query, searchParams.options?.resultGrouping, searchParams.options);
+          const results = await runSearch(db, searchParams);
           lastSearchResults = results;
           await updatePanelResults(searchPanel, results, searchParams.query, searchParams.options);
         }
@@ -1053,7 +1034,7 @@ async function replaceTagAll(
     // update the search panel
     await updatePanelQuery(searchPanel, searchParams.query, searchParams.filter);
     await updatePanelTagData(searchPanel, db);
-    const results = await runSearch(db, searchParams.query, searchParams.options?.resultGrouping, searchParams.options);
+    const results = await runSearch(db, searchParams);
     
     // Apply sorting to maintain current sort order
     const resultSettings = await getResultSettings();
@@ -1097,7 +1078,7 @@ async function removeTagAll(
   
   // update the search panel
   await updatePanelTagData(searchPanel, db);
-  const results = await runSearch(db, searchParams.query, searchParams.options?.resultGrouping, searchParams.options);
+  const results = await runSearch(db, searchParams);
   
   // Apply sorting to maintain current sort order
   const resultSettings = await getResultSettings();
